requirements.txt
tiktoken
pypdf
langchain_community
pinecone
langchain-pinecone 
pinecone-notebooks
python-dotenv
langchain-groq
